AI part 1

+++ RiD_RPGproject/RiD_RPGproject/Camera.cpp
@@ -7,7 +7,7 @@ void MP::Camera::_zoom_in()
 	if (_game_view.getSize().x > _min_x_size)
 	{
 		_game_view = _window.getView();
-		_game_view.zoom(0.8);
+		_game_view.zoom(float(0.8));
 		_window.setView(_game_view);
 	}
 
@@ -18,7 +18,7 @@ void MP::Camera::_zoom_out()
 	if (_game_view.getSize().x < _max_x_size)
 	{
 		_game_view = _window.getView();
-		_game_view.zoom(1.2);
+		_game_view.zoom(float(1.2));
 		_window.setView(_game_view);
 	}
 
@@ -28,8 +28,8 @@ MP::Camera::Camera()
 {
 	_min_x_size = RiD::ConfigurationLoader::getIntData("video settings", "minCameraView");
 	_max_x_size = RiD::ConfigurationLoader::getIntData("video settings", "maxCameraView");
-	_game_view.setSize(RiD::ConfigurationLoader::getIntData("video settings", "screenWidth")*0.78125, RiD::ConfigurationLoader::getIntData("video settings", "screenHeight"));
-	_gui_view.setSize(RiD::ConfigurationLoader::getIntData("video settings", "screenWidth"), RiD::ConfigurationLoader::getIntData("video settings", "screenHeight"));
+	_game_view.setSize(float(RiD::ConfigurationLoader::getIntData("video settings", "screenWidth")*0.78125), float(RiD::ConfigurationLoader::getIntData("video settings", "screenHeight")));
+	_gui_view.setSize(float(RiD::ConfigurationLoader::getIntData("video settings", "screenWidth")), float(RiD::ConfigurationLoader::getIntData("video settings", "screenHeight")));
 
 	_game_view.setViewport(sf::FloatRect(0, 0, 0.78125, 1)); //Computer will not drawing frames behind gui
 	_gui_view.setViewport(sf::FloatRect(0, 0, 1, 1));
